django_sso/requirements.txt
python-social-auth==0.3.6
social-auth-app-django==5.4.3
